Remove codec for inject/extract
In preparation for OT 0.32 support.
