--- dd-trace-ot/src/main/java/datadog/opentracing/propagation/Codec.java
@@ -1,46 +1,0 @@
-/*
- * Copyright (c) 2016, Uber Technologies, Inc
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-package datadog.opentracing.propagation;
-
-import datadog.opentracing.DDSpanContext;
-import io.opentracing.SpanContext;
-
-/** A codec is a simple object that can encode and decode a span context through a carrier */
-public interface Codec<T> {
-
-  /**
-   * Serialize the span context using the provided carrier
-   *
-   * @param context
-   * @param carrier
-   */
-  void inject(DDSpanContext context, T carrier);
-
-  /**
-   * Given a carrier, retrieve (rebuild) a span context. This context built will be use as the
-   * parent
-   *
-   * @param carrier
-   * @return the span context
-   */
-  SpanContext extract(T carrier);
-}
